- Added Logic To Adjust The Fill Factor Based On The Type Of Data
This change was made since a fill factor as low as 80 (the default) is not necessary for some columns since data is being inserted in the middle of the data less often.

Also added logic to filter out heaps.

This logic was written by Jeff McClure (@LPMcDude) in SQL and then translated to C# by @Jon-Corey.

Co-Authored-By: LPMcDude <98917940+LPMcDude@users.noreply.github.com>