- Made Internal and Public Naming Match
Renamed the Mental Health controls and attribute keys so that they are consistent with the public names for these questions.

Replaced "Psycho" (which was abbreviated from "Psychological" without thinking the abbreviation through) with "Mental Health" to match the label on the questions.

This is one of the reasons why abbreviating words should usually be avoided in code. It can lead to confusion and/or conveying the wrong meaning.