- Added Logic to Filter Out Coach Groups
Added logic to only show groups of group types that are tagged with the "Serve Toolbox" or "Groups Toolbox" tag. This is so that coach groups are not shown on the Leader Tools page.

This logic mimics the logic of the upcoming Group Leader Toolbox V2 updated that is planned for Q1 2025 and is on the feature-jc-GroupLeaderToolboxV2 branch.