- Added a Temporary Fix For The Live Metrics Block
Temporary fix for issue #6157. This change should be removed and replaced with the fix for issue #6157 when a fix is available.

This "fixes" the issue by commenting out the code that was causing performance issues. This "fix" assumes that we're not using a SoftRoomThreshold on parent locations (which we aren't). If we are, the count will be off on parent locations in this block.