- Fixed Translation System And Additional Registrar Fields
Fixed the Registration Attribute options not translating. Copied the logic for translating options from the Registrant form and then adjusted the logic to fit better into this particular form.

Added the missing Preferred Language field to the useLoggedInPersonForRegistrar registrar form.

Added logic to update the Phone Number and Preferred Language when the values in the controls are edited.