- Fixed Timeout When Using an Invalid Document Type ID
Fixes an issue where when creating a new signature document or using a non-integer parameter this block would try to load all of the signature documents, which would cause a timeout.

This fixes that issue by not running the query if the provided Document Type ID is either 0 or null (non-integer parameters are converted to null).